backend/requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
python-dotenv>=0.19.0,<0.20.0
httpx>=0.23.0
openai>=1.0.0
sqlalchemy>=1.4.0,<1.5.0
psycopg2-binary>=2.9.0,<2.10.0
email-validator>=1.1.0,<1.2.0
python-jose[cryptography]>=3.3.0,<3.4.0
passlib[bcrypt]>=1.7.0,<1.8.0
python-multipart>=0.0.5,<0.0.6
SQLAlchemy>=1.4
